DRY list_elements!
