+++ helixflow/tests/slint_surreal.rs
@@ -13,7 +13,9 @@ fn test_slint_with_surreal() {
     prepare_slint!();
 
     let backend = Rc::new(SurrealDb::new().unwrap());
+    
     let helixflow = HelixFlow::new().unwrap();
+    list_elements!(&helixflow);
 
     let hf = helixflow.as_weak();
     let be = Rc::downgrade(&backend);
